requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.5
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
tzlocal==5.2
urllib3==2.0.4
yarl==1.9.2

SQLAlchemy~=2.0.23